portability on Linux

+++ mddstate.hpp
@@ -11,6 +11,7 @@
 
 #include "intvar.hpp"
 #include <set>
+#include <cstring>
 
 
 /*
@@ -68,14 +69,14 @@ public:
    int at(int i) const { return _state[i];}
    int operator[](int i) const { return _state[i];}  // to _read_ a state property
    void set(int i,int val) { _state[i] = val;}       // sets a state property 
-   long hash() { 
-      long ttl = 0;
+   long long hash() { 
+      long long ttl = 0;
       for(auto v : _state)
          ttl = (ttl << 8) + (ttl >> (64 - 8)) + v;
       _hash = ttl;
       return _hash;
    }
-   long getHash() const noexcept { return _hash;}
+   long long getHash() const noexcept { return _hash;}
    bool operator==(const MDDState& s) const {    // equality test likely O(1) when different. 
       if (_hash == s._hash) {
          bool eq = true;
